work on the live image crosssection
