+++ src/bluesky/callbacks/mpl_package/limit_factories.py
@@ -1,0 +1,97 @@
+import logging
+
+import numpy as np
+
+logger = logging.getLogger(__name__)
+
+
+def fullrange_limit_factory(limit_args=None):
+    """
+    Factory for returning full-range limit functions
+
+    limit_args is ignored.
+    """
+
+    def _full_range(im):
+        """
+        Plot the entire range of the image
+
+        Parameters
+        ----------
+        im : ndarray
+           image data, nominally 2D
+
+        limit_args : object
+           Ignored, here to match signature with other
+           limit functions
+
+        Returns
+        -------
+        climits : tuple
+           length 2 tuple to be passed to `im.clim(...)` to
+           set the color limits of a ColorMappable object.
+        """
+        return (np.nanmin(im), np.nanmax(im))
+
+    return _full_range
+
+
+def absolute_limit_factory(limit_args):
+    """
+    Factory for making absolute limit functions
+    """
+
+    def _absolute_limit(im):
+        """
+        Plot the image based on the min/max values in limit_args
+
+        This function is a no-op and just return the input limit_args.
+
+        Parameters
+        ----------
+        im : ndarray
+            image data.  Ignored in this method
+
+        limit_args : array
+           (min_value, max_value)  Values are in absolute units
+           of the image.
+
+        Returns
+        -------
+        climits : tuple
+           length 2 tuple to be passed to `im.clim(...)` to
+           set the color limits of a ColorMappable object.
+
+        """
+        return limit_args
+
+    return _absolute_limit
+
+
+def percentile_limit_factory(limit_args):
+    """
+    Factory to return a percentile limit function
+    """
+
+    def _percentile_limit(im):
+        """
+        Sets limits based on percentile.
+
+        Parameters
+        ----------
+        im : ndarray
+            image data
+
+        limit_args : tuple of floats in [0, 100]
+            upper and lower percetile values
+
+        Returns
+        -------
+        climits : tuple
+           length 2 tuple to be passed to `im.clim(...)` to
+           set the color limits of a ColorMappable object.
+
+        """
+        return np.percentile(im, limit_args)
+
+    return _percentile_limit
